(Ada) fix namequal. bug f. package renaming itself
A package renaming itself produced an endless loop when it renamed
itself. The cause was that the package was not visible, so it chose its
renaming, but that was declared itself within the package -> endless
loop.

Solution is to add all named open scopes to the visible scopes, so they
take priority over renamed scopes.

RC-1176