(Ada) Revise secondary variable initalizations
In Ada multiple variables and parameters can be (default) initialized with a single initializer.
a, b: Integer := InitExpr;

The ROSE AST and the C language do not provide for that. Thus far, such
initialization had been represented as:
int a = InitExpr, int b = clone(InitExpr).

This code could suffer from side-effects if InitExpr was not side-effect
free.

To mitigate this problem, at least for variables TV suggested to use the
following representation for variables.
int a = InitExpr, int b = a

The representation of parameter initializers remains unchanged.

RC-1301