(Ada) revise parameter list generation
In the Asis representation, parameter references inside of routine
bodies refer back to the parameter list of the first declaration.

Not sure if this is by design or a bug. As work around, use the first
decl's parameter list to generate the body's parameter list. Unlike C++,
Ada requires parameter names AND default intializers of the declaration
and definition to match.

One caveat is that multiple initialized names could
have a combined VariableDeclaration in the declaration, while having
separate VariableDeclarations in the definition (and vice versa).
This workaround does not preserve this distinction in the AST.

RC-1372