(Ada) parameter references in routines
Asis refers back to the parameters in the first declaration.
This seems incorrect (at least in the context of ROSE).
To fix this, a previous commit cloned the parameter list of the original
decl and made it available to the routine body.

This failed, when a separated body was used, since it did not refer back
to the first declaration, but onlt to the procedure body stub.

Fixed by reading through the body stub to find the first declaration.

+Added regression test.

RC-1418