(Ada) remove unnecessary try begin-end
ROSE represents Ada blocks with exception handlers

begin
 ...
exception
  when ... => ...
end;

in the following way:

{ // outer block
  try { ... }
  catch (...) { ... }
}

if the try block is the only statement in the block, the generation of
the begin/end of the outer block produces code that does not compile.

Solution: do not generate begin/end of try statements, when the
statement is the only statement in an explicit block.

RC-1432