(Ada) pragma placement
- Improves pragma placement by no longer considering compiler generated
nodes for threading the pragmas into place (into AST node sequences).

- Adds heuristic to add a pragma between declaration scope and
  statement scope to the declaration scope (needed for customer code).

Generally, the problem is difficult to solve:
Asis maintains a separate list of pragmas for scopes. The pragmas have
to be threaded into place of AST statements and declaration lists based
on their source code location.

The fact that one pragma list can be associated with declaration and
statement scope complicates the problem.
Consider:
decl1;
decl2;
pragma A
begin -- not explicitely represented in ASIS
  pragma B;
  null;
end;

How do we know that pragma A belongs to the declaration scope and pragma
B to the statement scope?

A possible solution is to use the ASIS information that ties pragmas to
declarations...

For now the test case, pragma_placement.adb illustrating difficulties of
finding the appropriate place for a pragma has been added to the test
cases (will be tracked under JACCEL-343).

RC-1432