(Ada) distinguish between separate and child units
to know whether a unit is a separate proc or child unit proc, the
preceeding declaration needs to be checked. to this end the body scope
was traversed and the first function that had the same first_nondefining
declaration set was checked for having the separate modifier.

instead of checking the first function, the scope can be traveresed
backward to check the LAST function with the same first_nondefining
link.

RC-1477