(Ada) compute source location when not in ASIS
- changed computation of source location when the information is not in
ASIS.

e.g., for try statement (which is compiler-generated to match a C++
AST), the catch handlers, and the try block.

When the source information is not present, it will be derived from the
children during AST construction.

All other nodes which do not have a source location will get assigned
one in the AST post processing step of the Ada translator (this has been
done for some time).

RC-1484