(Fortran Flang) Added SgFortranContinueStmt to ROSETTA
The Fortran CONTINUE statement has different semantics in C/C++ than Fortran.
In Fortran continue is a noop.  The Fortran CYCLE statement has similar
semantics to the C/C++ continue statement and the use of SgContinueStmt for
Fortran CYCLE has not changed.  The old Fortran parser implementation uses
an SgLabelStatement for CONTINUE?  This should be updated to use SgFortranContinueStmt.

GL-6