(ByteCode JVM) Configure with --with-alloc-memset=3
The call to Rose::AST::IO::free() in testTranslator fails because the JVM
constant pool is not freed.  This commit _starts_ the process of consistently
freeing the JVM AST nodes.

However, the design of freeing binary AST nodes appears to be inconsistent with
--with-alloc-memset=3 because the AST has many checks like

    if (node != nullptr) {
       delete node;
    }

This doesn't work with --with-alloc-memset=3 because deleted nodes are set to
a sentinel value other than NULL.

Rose::AST::IO::free() needs to be rethought for binary analysis.

GL-295