(Ada) expose implicit dereferences in IR
this includes the following contexts:
+ array access
+ subroutine call (*1)
+ field access
+ certain Ada attributes (*2)

The respective AdaMaker functions were extended to insert the
dereferences during the initial AST construction.

Type computation (such as getArrayInfo and typeOfExpr) in
SageInterfaceAda were reorganized to better deal with
dereferences and array indexing in general. Specifically, this
functions were extended to deal with SgAdaFormalTypes.

*1 In C++, the subroutine calls through pointers are currently
   not resolved (though they can appear if explicitely used in code).
   For consistency with the rest of the Ada representation, the deref
   operation nodes are added in Ada.

*2 The implementation is incomplete for some attribute names..
   notable 'Valid, because it is unclear how to test for the
   property. Since an access type does not have 'Valid, it is
   assumed that the context requires deref.

RC-1556