(ByteCode) Implement CFG
Major improvements in implementation of CFG for CIL
- required refactoring for JVM

Added std::shared_ptr(s) to:
- ByteCode::Container member "std::vector<std::shared_ptr<Namespace>> namespaces_;"
- ByteCode::Namespace member "std::vector<std::shared_ptr<Class>> classes_;"

Added virtual functions:
- isSystemReserved(name) for, e.g., "Microsoft" or "java/"
- ByteCode::Class::typeSeparator() returning "." or "::"

- removed virtual void destructorHelper() from src/AstNodes/BinaryAnalysis/SgAsmGenericHeader.h

Implemented new method get_MethodList_object(const SgAsmCilMethodDef*) to SgAsmCilTypeDef.
This is needed so that the MethodList object can be found that corresponds to the
SgAsmCilMethodDef* parameter. See the CIL specification to see how the MethodList object
works. It is a bit strange, probably to minimize executable code size.

RC-ROSE-2