(Ada) model Character types as enum
following the Ada standard, character types are now modeled as enum.
This allows character to be used, such as in:
  use type Character;

Note, link to a base representation is through the declarations
adaParentType which links to SgTypeChar, SgTypeChar16, SgTypeChar32
for Character, Wide_Character, and Wide_Wide_Character respectively.

Inherited enumerators are not injected into the enum declarations,
because this would be prohibiteve for the Wide_*Character types.

TODO: unparsing use type at the unit scope:
  use type Standard.Character

RC-1595