(Binary Analysis) New alignment class
When aligning a value upward, most of ROSE doesn't check that the
operation was successful (didn't overflow). Doing so is a PITA that
requires a temporary variable, like this:

    const auto temp = alignUp(x, something.alignment);
    if (temp < x || temp > Rose::BitOps::lowMask<uint64_t>(something.addressSizeInBits))
        handle_overflow();
    x = temp;

This commit adds `Rose::BinaryAnalysis::Alignment` that holds the
alignment and the size in bits and the align-up operation returns an
optional value.

If an exception is okay, do this:

    x = *something.alignment.alignUp(x);

If you want some default value, do this:

    x = something.alignment.alignUp(x).orElse(1);

If you want to handle the overflow explicitly, do this:

    if (!something.alignment.alignUp(x).assignTo(x))
        handle_overflow();

Issue #84