(ByteCode) Separate parsing from SgProject creation
The frontend() function previously called a SgProject constructor that,
not only created and initialized the project object, but also parsed
all files on the command line to create the AST. This is bad object design.

Fixed by using the default constructor to create and initialize a project,
then calling parse() within frontend().

This modification allows AST construction to be done by an agent other than
SgProject, e.g., a Rose::BinaryAnalysis::Partitioner2::Engine

The existing SgProject constructors (that call parse) should be deprecated
(not done in this commit).

JVM-1
LDRD_SI-1