(Jovial) SegFault related to comments
Problem was:
  SageInterface::setBaseTypeDefiningDeclaration was attempting to move comments eventually
  calling SageInterface::removeStatement of the statement containing the comment. Problem
  arose when the comment location was before the statement and statement was the first
  statement in the block.  Problem being a segfault.

Consideration should be given to moving comments to the variable declaration in setBaseTypeDefiningDeclaration,
first, to circumvent (by artfulness or deception) the segfault. The variable declaration is a better location
for the comment anyway, rather than picking some other statement, before or after, to attach the comment.

Fixed by moving before comments to the variable declaration in SageTreeBuilder, not SageInterface.
SageInterface not meaningfully changed.

GL-349