Add Jovial filenames to SourceFileSuffixList
It appears that Jovial, Ada, and Java .class files are processed by the frontend because of the binary exception, meaning that no suffix is a valid and should be processed, see validSourceFileSuffixes.push_back("").

This is confusing at cursory glance of ROSE's frontend. Also doesn't seem to be a general solution and probably wasn't designed for this behavior.

Issue #100