(Jvm) Implement jar file handling
Added ModulesJvm.{h,C} to manage jar/zip files.

Includes a copy of much of the open source project, zip_file.cpp. The license of
this file is based on the mit-license and should be no problem to use.

ModulesJvm only uses this library for decompression of class files
from a jar file that it has opened. This part of the library (decompression)
is based on a subset of zlib. However, the software from zip_file.cpp has not
been trimmed to include only the necessary decompression routines. I wanted to get
something committed that works and then start to whittle back.

JVM-1
LDRD_SI-1