(Binary Analysis) Fixed sysv calling convention; improve debug output
The SYSV calling convention for AMD64 didn't define the registers for
the initial integer or floating-point arguments, and therefore always
allocated arguments on the stack.

Also added non-null register dictionaries to the `ConcreteLocation`
objects that are used in the calling conventions. This improves the
debugging output by showing register names rather than just the
descriptor quads.

Issue #132