(check-tools) Updated Codethorn testing scripts to run in any directory
codethorn has a bunch of testing scripts that call codethorn, but it
assumed that codethorn was in the current directory. This commit allows
you to pass in a directory where codethorn is located. It follows the
argument style for each script individually (some have a switch, some
are positional)

RC-1630