(Engineering) Accept additional Gitlab issue references in commit mesgs
The following Gitlab issue reference styles are accepted in addition
to those that were already accepted:

* Phrases like "Issue #217" and "Issue #217+" which mean issue 217 in
  the Gitlab group to which this repository belongs. Gitlab
  understands the "#217" and "#217+" part as an issue link; the word
  "Issue" before it is because Git will treat lines that start with a
  hash as if they are comments.

* Phrases like "rose#217" and "rose#217+" which mean issue 217 in a
  project named "rose" within Gitlab. Gitlab understands the entire
  phrase.

* Phrases like "foo/testing#217" and "foo/testing#217+" which mean
  issue 217 in a project named "foo/testing". Gitlab understands the
  entire phrase.

The versions that end with a "+" cause Gitlab to insert the issue
title as the link text.

The policy checker also makes sure that the issue reference list is
the last thing in the commit message.

Issue #217
Issue #217+
ROSE#217
ROSE#217+
ROSE-Compiler/ROSE#217
ROSE-Compiler/ROSE#217