(Jovial) Enum on partial include
Import of a CompoolModule type but not the status variable.

Fixed by setting the type of the SgAtOp lhs operand (was SgTypeUnknown). When a variable
is declared in an imported CompoolModule, an alias to the variable name may not be imported.
Now the symbol is found, if needed, by searching for the name in the scope of
the CompoolModule where the type was defined.

Issue #405