(Binary Analysis) All smart ptrs use same dynamic casting function
Previously, different pointer types (STL, Boost, Sawyer, raw) had
different ways to do a dynamic cast. Although ROSE already often hides
the pointer implementations behind "Ptr" types, the user still had to
know the implementation in order to know how to do a dynamic cast.

This commit fixes that by replacing the variety of dynamic casts with
a single construct: `Rose::as<T>(U)` which converts a pointer to `U`
to a pointer to `T` no matter the implementation.

This commit does NOT change Sage AST node down-casts from `isSgFoo(x)`
to `as<SgFoo>(x)`, although `as` can also be used for that purpose.

Issue #400