(BinaryAnalysis) Ability to download data from a VxWorks terminal
This commit adds the ability to load data into ROSE's `MemoryMap` by
connecting to a VxWorks telnet server over TCP/IP. As such, it
replaces a previous dedicated download tool and improves it in a
number of ways:

* The new "vxworks:..." resource string is understood by all ROSE
  binary analysis tools. There is no longer any need to download
  data into a temporary raw binary file when using these tools.

* The resource string specifies a number of settings that were
  previously only adjustable by modifying the downloader source
  code. These include the prompt string, the prompt delay, and
  data transfer size.

* The resource string incorporates a number of settings that were
  separate command-line arguments in the previous downloader tool,
  including starting address and size, host IP address or name,
  and TCP/IP port number.

* The new mechanism does not require the telnet client. It makes
  connections directly to the server.

* Since the existing bat-mem tool now understands VxWorks
  terminals (see first bullet) and can already dump a `MemoryMap`
  in various formats (including binary blobs), it fully replaces
  the dedicated dumper tool.

* The new mechanism is more resilient to errors and has much more
  consistency checking. It should be much more difficult to download
  incorrect data without some sort of warning from the tools.

* The new mechanism is fully documented in the "--help" output of all
  binary analysis tools.

* The new mechanism is not constrained to operate on data regions
  whose starting address and sizes are 16-byte aligned.

* The new mechanism is more extensible so we can more easily write
  ROSE-based tools that do other things besides download data over the
  VxWorks terminal.

This commit also adds a rudimentary telnet server for testing. The
server emulates the VxWorks "d" (memory dump) command and other
aspects of the terminal. Being a normal ROSE binary analysis tool, it
can initialize the simulated memory with any binary specimen,
including data downloaded from a VxWorks terminal. This ability
enables this tool act as a fast local cache of a slow VxWorks
connection since the fake server is not restricted by VxWorks
limitations in data transfer sizes or command timings.

Issue #466