(Jovial) Frontend crash on comment
This issue was likely caused by changes made to allow parsing of files
with unprintable characters in them (issue gl-439). These changes
required changing the lexer syntax to fix the specification of
the ascii unprintable character codes producing LAYOUT. The ascii
character values should be specified in decimal, not octal as was
done for issue 439. Corrected here.

Issue #439
Issue #452