(Model Checker) SemanticCallbacks keep track of the stack memory region
The SemanticCallbacks for the PartitionerModel keeps track of the
region of memory that has been allocated to the stack instead of
trying to guess it later.

Moved some constructor assertion checks earlier using `notnull`.

Replaced a `dynamic_cast` with `as`.

Partitioner model RISC operators can return a pointer to the semantic
callbacks that were provided when the RISC operators object was
constructed.

Issue #503