(cleanup) removed old printf debugging statement
It looks like Dan was trying to figure something out a while ago, then
figured it out and added the case to the ROSE_ASSERT without removing
the printf statement in its original form, so now the old failing case
just emits a random debugging message. This removes it.

Issue #521