(refactor) pass global var as parameter
In commandline_processing.h/C: remove the need to access
Rose::is_Ada_language directly, and pass it in as argument.

Note, is_Ada_language would be set through -x on the command line.
In this case, it enables the recognition of .a files as Ada files.
If not set, .a files are treated as binary objects and passed to the
backend (source processin) or binary analysis.

Updated files that directly and indircetly call isAdaFileNameSuffix.
Did not fix the issue in the ROSE GUI, b/c that seems mostly obsolete.

Issue #608