(Jovial) Partial include not grabbing table item
The variable CON is not added as part of the compool directive. This can be fixed by just
doing a full include but the partial include code should be improved as well.

The problem is due to a major ambiguity in the grammar related to function calls. The
grammar constructs UserDefinedFunctionCall but must be disambiguated as a variable
declaration of a table item from the accompanying COMPOOL file. This can only be done
if all of the symbols are loaded from the file. Checking for the case of

    if (auto modifier_type = isSgModifierType(type)) --> then loadTypeSymbol

loads the variable symbols from the table declaration in the COMPOOL file.

Issue #478