(Jovial) Status shadowed on procedure call
StatusConstants are transformed to an SgEnumVal. The symbol for the enum
value is obtained from the symbol table.  However, since enums are globally
scoped there can be multiple symbols that should be matched with the correct type.

This commit fixes part of the problem by creating a new SgEnumVal rather
than sharing the enum value obtained from the symbol table. The SgEnumVal
is new (same field name and value) but shares the declaration of the original.

Issue #476