(Jovial) Status shadowed on procedure call
StatusConstants are transformed to an SgEnumVal. The symbol for the enum
value is obtained from the symbol table.  However, since enums are globally
scoped there can be multiple symbols that should be matched with the correct type.

This commit adds the function matchExprType() to convert the type of an expression
to that of the given type if possible and uses it to match the types in a
BinaryExpression (Note, there will likely be other productions to convert).

Issue #476