(Jovial) Status shadowed on procedure call
StatusConstants are transformed to an SgEnumVal. The symbol for the enum
value is obtained from the symbol table.  However, since enums are globally
scoped there can be multiple symbols that should be matched with the correct type.

Issue completed by converting all actual arguments in a function call to match the
type of the formal parameters.

Issue #476