(Engineering) Fixed compiler warnings
Most of these were `ASSERT_not_null(this)` which were rewritten to
`ASSERT_this()`. It is C++ UB for `this` to be null and LLVM issues
warnings about checking whether `this` is null or not null. This the
warning comes from the macro definition, they're all blamed on the
same person.

The `ASSERT_this` macro is a no-op in ROSE since it is UB for `this` to
be null.

Issue #6