(Engineering) Attempt some early short-circuiting of concepts
This commit attempts to perform some early short-circuiting of
concept support, in an attempt to ignore concepts in header files
for the sake of getting C++20 and C++23 code tested with Plumhall.

The main change right now is to detect templk_concept and return an
SgEmptyDeclaration. However, more issues are popping up which is
preventing actual GCC 13 headers to get parsed.

Issue #711