(ROSE) Fixing performance of ROSE on large code files
Added optional behavior to ignore name qualification of the
translation unit outside of the source file:
   -rose:suppressNameQualificationAcrossWholeTranslationUnit
This improved the performance by a factor of 86x for the name
qualification. Also, fixed redundant conversion of C++ string
to C-style string to C++ string in the hash function for the
symbol table support (however, this didn't appear to make any
difference to the performance).

Issue #707