(Headers) Replace GNU 13 type_traits header with ROSE specific version
This commit replaces the GNU 13 type_traits header with a ROSE
specific version. This version of the type_traits header has the use
of concepts unguarded by the __cpp_concepts feature macro when C++23
is enabled. ROSE currently sets __cpp_concepts to 0 because concepts
are not yet supported, and in all other header files this enables
non-concept code, except for in this specific header.

Issue #752