(Engineering) Comment out __is_identifier check because of EDG bug
This commit comments out preprocessor code which uses __is_identifier
due to an EDG bug. EDG implements __is_identifier as a feature macro
which decides at compile-time whether or not a token is an identifier.
This can be used to check if a type is defined natively in Clang. For
instance, __is_identifier(abc) would return 1, but __is_identifier(float)
would return 0. This can be used at preprocessing time to determine that
float is a native type, but abc is not. In my Clang, tests,
__is_identifier(_Float16) returns 0, telling the preprocessor that
_Float16 is a native type supported by the compiler. However, even when
_Float16 is a legal native type in EDG 6.5, __is_identifier(_Float16)
returns 1 erroneously, misidentifying it to the preprocessor as an
identifier, and not a native type. Therefore, we cannot reliably use
__is_identifier to identify a token as a native type with EDG 6.5,
at least when it comes to extended floating-point types.

Issue #752