(Engineering) Capture version-specific macros for each C/C++ version
This commit changes the behavior of capturing macros from the host
compiler. Previously, predefined macros from the host compiler were
captured using the default C or C++ version, and this list of macros
was used statically regardless of the action C or C++ version we use
when invoking rose-compiler. The new behavior instead captures the
list of predefined macros for every C and C++ version as of
September 26, 2025, and dumps them to a bunch of separate header files.
These are then conditionally included in rose_edg_required_macros_and_functions.h
depending on what version was set when invoking rose-compiler. This ensures
that only the correct predefined macros are included for each individual
C or C++ version.

Issue #765