(Engineering) Change GNU default standard from gnu11/gnu++11
This change alters a previous commit setting the default standard
when a GNU compiler backend is used. Previously, gnu11/gnu++11 was
set, but this has been changed in light of issues in the pipeline
where a default standard of C11/C++11 was unfavorable, but retains
setting the GNU features. The behavior before the initial change
should be maintained, with the caveat that now we enforce GNU extension
support.

Issue #765