(C++23) Fix C++23 support with new compiler-specific macro functionality
This commit fixes C++23 support after a prior merge request changed C and
C++ parsing to use the correct compiler-specific macros. There are a couple
of macros defined in C++23 that should be disabled because they cause issues
due to lacking EDG support. This has been fixed with this merge request.

Issue #752