(CMake) Remove unnecessary explicit OBJECT dependency
CMakeLists.txt files in subdirectories of "src" don't need to
explicitly specify target_include_directories or target_link_libraries
because this is done in "src/CMakeLists.txt" for all OBJECT libraries
through the rose_object_library_dependencies INTRERFACE library, which
is a dependency of every OBJECT library that goes into creating
librose.so.

This is a partial revert of commit 2322b134.

Issue #888