(BinaryAnalysis) Refactor SerialIo into frames and backends
This refactoring splits out the object framing and file serialization logic from SerialIo into dedicated framing and serialization backend classes.
The intention is to move messy conditional compilation logic in SerialIo into files that are wholly enabled/disabled at configure time,
and also to enable a clean switch between Flatbuffer and Boost serialization.

The main breaking change is that some types and classes that were part of the SerialIo class are now part of a new Serialization namespace.
The most notable migrated enum is SerialIo::Format.

This is all motivated by Issue #898 and an eventual migration from Boost serialization to Flatbuffer serialization.